# js/frontends/system-docs/openapi/organizer-api.yml
openapi: 3.0.0
paths:
...
  /admin/users/{id}:
    get:
      operationId: AdminUsersController findOne
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: number
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: ''
          content:
            application/json:
              schema:
                example:
                  statusCode: 404
                  message: User with id 1 not found
                  error: Not Found
      tags: *ref_2
    patch:
      operationId: AdminUsersController update
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAdminUserDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: ''
          content:
            application/json:
              schema:
                example:
                  statusCode: 404
                  message: User with id 1 not found
                  error: Not Found
      tags: *ref_2
    delete:
      operationId: AdminUsersController remove
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: number
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: ''
          content:
            application/json:
              schema:
                example:
                  statusCode: 404
                  message: User with id 1 not found
                  error: Not Found
      tags: *ref_2
...
  /agency/users/{id}:
    get:
      operationId: UsersController findOne
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: number
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: ''
          content:
            application/json:
              schema:
                example:
                  statusCode: 404
                  message: User with id 1 not found in agency 1
                  error: Not Found
      tags: *ref_5
    patch:
      operationId: UsersController update
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: ''
          content:
            application/json:
              schema:
                example:
                  statusCode: 404
                  message: User with id 1 not found in agency 1
                  error: Not Found
      tags: *ref_5
    delete:
      operationId: UsersController remove
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: number
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: ''
          content:
            application/json:
              schema:
                example:
                  statusCode: 404
                  message: User with id 1 not found in agency 1
                  error: Not Found
      tags: *ref_5
  /agency/users:
    get:
      operationId: UsersController findAll
      parameters: [ ]
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
      tags: *ref_5
    post:
      operationId: UsersController create
      parameters: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserDto'
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: ''
          content:
            application/json:
              schema:
                example:
                  statusCode: 400
                  message: Email is already taken
                  error: Bad Request
      tags: *ref_5
...
info:
  title: Organizer API
  description: ''
  version: '1.0'
  contact: { }
...
servers: [ ]
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: number
          example: 1
        email:
          type: string
          example: jan@kowalski.com
        firstName:
          type: string
          example: Jan
        lastName:
          type: string
          example: Kowalski
        phoneNumber:
          type: string
          example: '+48123456789'
        timeCreated:
          format: date-time
          type: string
          example: '2022-11-26T18:47:02.541Z'
        timeSignOut:
          format: date-time
          type: string
          example: '2022-11-26T18:47:02.541Z'
        role:
          enum:
            - 0
            - 1
            - 2
            - 3
            - 4
          type: number
          example: 2
        agencyId:
          type: number
          example: 1
      required:
        - id
        - email
        - firstName
        - lastName
        - phoneNumber
        - timeCreated
        - timeSignOut
        - role
    CreateUserDto:
      type: object
      properties:
        email:
          type: string
          maxLength: 320
          example: jan@kowalski.com
        password:
          type: string
          minLength: 16
          maxLength: 128
          example: '&Y+sFaS{&d>8ycO)FLhF41qiQk{IYEb?'
        confirmPassword:
          type: string
          example: '&Y+sFaS{&d>8ycO)FLhF41qiQk{IYEb?'
        firstName:
          type: string
          maxLength: 100
          example: Jan
        lastName:
          type: string
          maxLength: 100
          example: Kowalski
        phoneNumber:
          type: string
          maxLength: 30
          example: '+48123456789'
        role:
          enum:
            - 0
            - 1
            - 2
            - 3
            - 4
          type: number
          example: 2
      required:
        - email
        - password
        - confirmPassword
        - firstName
        - lastName
        - phoneNumber
        - role
    UpdateAdminUserDto:
      type: object
      properties:
        firstName:
          type: string
          maxLength: 100
          example: Jan
        lastName:
          type: string
          maxLength: 100
          example: Kowalski
        phoneNumber:
          type: string
          maxLength: 30
          example: '+48123456789'
...
    UpdateUserDto:
      type: object
      properties:
        firstName:
          type: string
          maxLength: 100
          example: Jan
        lastName:
          type: string
          maxLength: 100
          example: Kowalski
        phoneNumber:
          type: string
          maxLength: 30
          example: '+48123456789'
        role:
          enum:
            - 0
            - 1
            - 2
            - 3
            - 4
          type: number
          example: 2
...